+++ src/wetwire_gitlab/importer/__init__.py
@@ -1,5 +1,6 @@
 """YAML importer module for wetwire-gitlab."""
 
+from .codegen import generate_python_code
 from .ir import IRInclude, IRJob, IRPipeline, IRRule
 from .parser import parse_gitlab_ci, parse_gitlab_ci_file
 
@@ -8,6 +9,7 @@ __all__ = [
     "IRJob",
     "IRPipeline",
     "IRRule",
+    "generate_python_code",
     "parse_gitlab_ci",
     "parse_gitlab_ci_file",
 ]